third_party/WebKit/LayoutTests/fast/html/imports/rel-import-to-style-expected.txt

Updating attribute value shouldn't invalidate loaded import.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Before update:
PASS target.sheet is null
PASS target.import is non-null.
After update:
PASS target.sheet is null
PASS target.import is non-null.
PASS successfullyParsed is true

TEST COMPLETE
